Move all disBatch logic into disbatch module, rather than main script
